Changes for Repeat Command
Had to add a getHistoryCommand to retrieve the history command.

Moved handling of the inputed line to it's own function so the repeat
function could call the same function from within itself. This will not
cause infinite recursion at the moment because there are no other
recursion possibilities and repeat commands are not added to the
history.